Exercises 4.15-4.22 done and part4 is completed.

+++ part4/bloglist-backend/controllers/blogs.js
@@ -1,8 +1,11 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+        .find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs.map(blog => blog.toJSON()))
 })
 
@@ -22,6 +25,17 @@ blogsRouter.get('/:id', async (request, response, next) => {
 blogsRouter.post('/', async (request, response, next) => {
     const body = request.body
 
+    //const token = getTokenFrom(request)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+        return response.status(401).json({
+            error: 'token missing or invalid'
+        })
+    }
+    const user = await User.findById(decodedToken.id)
+
+    // const user = await User.findById(body.userId)
+
     if (!body.title || !body.url) {
         return response.status(400).json({ error: 'Title or URL is missing' })
     }
@@ -29,10 +43,14 @@ blogsRouter.post('/', async (request, response, next) => {
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes || 0
+        likes: body.likes || 0,
+        user: user._id
     })
     try {
         const savedBlog = await blog.save()
+        user.blogs = user.blogs.concat(savedBlog._id)
+        await user.save()
+
         return response.json(savedBlog.toJSON())
     } catch (exception) {
         next(exception)
@@ -40,9 +58,17 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (!request.token || decodedToken.id.toString() !== blog.user.toString()) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
     try {
         await Blog.findByIdAndRemove(request.params.id)
-        response.status(204).end()
+        return response.status(204).end()
     } catch (error) {
         next(error)
     }
@@ -50,16 +76,35 @@ blogsRouter.delete('/:id', async (request, response, next) => {
 
 blogsRouter.put('/:id', async (request, response, next) => {
     const body = request.body
+
     const id = await Blog.findById(request.params.id)
 
     if (!id) {
-        return response.status(404).json({ error: 'ID not found.' })
+        return response.status(404).json({ error: 'resource not found.' })
     }
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!request.token || !decodedToken.id) {
+        return response.status(401).json({
+            error: 'token missing or invalid'
+        })
+    }
+    const user = await User.findById(decodedToken.id)
+
+    if (!user) {
+        return response.status(401).json({ error: 'Not autherized' })
+    }
+    if (!body.title || !body.url) {
+        return response.status(400).json({ error: 'Title or URL is missing' })
+    }
+
     const blog = {
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes
+        likes: body.likes,
+        user: body.userId
     }
 
     try {
